refactor: migrate useDecision() to focused hooks + extract KeyboardShortcutsModal
- Migrate 10 components from monolithic useDecision() to focused hooks:
  useDecisionData(), useResultsContext(), useActions()
- Each component now subscribes only to the context slices it needs,
  reducing unnecessary re-renders
- Extract KeyboardShortcutsModal from page.tsx into standalone component
  with self-contained focus trap logic
- All 1502 unit tests pass

Components migrated:
  ImportModal      → useActions()
  CompareView      → useDecisionData()
  SensitivityView  → useResultsContext() + useActions()
  MonteCarloView   → useDecisionData() + useResultsContext()
  AHPWizard        → useDecisionData() + useActions()
  EnrichmentSuggest→ useDecisionData() + useActions()
  Header           → useDecisionData() + useActions()
  DecisionBuilder  → useDecisionData() + useActions()
  ResultsView      → useDecisionData() + useResultsContext() + useActions()
  page.tsx         → useDecisionData() + useActions()

Closes #200
